# colexec: add support for ILIKE and NOT ILIKE