--- a/pkg/sql/colexec/colexecprojconst/like_ops.go
+++ b/pkg/sql/colexec/colexecprojconst/like_ops.go
@@ -32,7 +32,7 @@ func GetLikeProjectionOperator(
 	pattern string,
 	negate bool,
 ) (colexecop.Operator, error) {
-	likeOpType, patterns, err := colexeccmp.GetLikeOperatorType(pattern, negate)
+	likeOpType, patterns, err := colexeccmp.GetLikeOperatorType(pattern)
 	if err != nil {
 		return nil, err
 	}
@@ -45,45 +45,21 @@ func GetLikeProjectionOperator(
 		outputIdx:      resultIdx,
 	}
 	switch likeOpType {
-	case colexeccmp.LikeConstant:
-		return &projEQBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        pat,
-		}, nil
-	case colexeccmp.LikeConstantNegate:
-		return &projNEBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        pat,
-		}, nil
-	case colexeccmp.LikeNeverMatch:
-		// Use an empty not-prefix operator to get correct NULL behavior.
-		return &projNotPrefixBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        []byte{},
-		}, nil
 	case colexeccmp.LikeAlwaysMatch:
 		// Use an empty prefix operator to get correct NULL behavior.
 		return &projPrefixBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        []byte{},
+			negate:          negate,
 		}, nil
-	case colexeccmp.LikeSuffix:
-		return &projSuffixBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        pat,
-		}, nil
-	case colexeccmp.LikeSuffixNegate:
-		return &projNotSuffixBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        pat,
-		}, nil
-	case colexeccmp.LikePrefix:
-		return &projPrefixBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        pat,
-		}, nil
-	case colexeccmp.LikePrefixNegate:
-		return &projNotPrefixBytesBytesConstOp{
+	case colexeccmp.LikeConstant:
+		if negate {
+			return &projNEBytesBytesConstOp{
+				projConstOpBase: base,
+				constArg:        pat,
+			}, nil
+		}
+		return &projEQBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        pat,
 		}, nil
@@ -91,21 +67,13 @@ func GetLikeProjectionOperator(
 		return &projContainsBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        pat,
+			negate:          negate,
 		}, nil
-	case colexeccmp.LikeContainsNegate:
-		return &projNotContainsBytesBytesConstOp{
+	case colexeccmp.LikePrefix:
+		return &projPrefixBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        pat,
-		}, nil
-	case colexeccmp.LikeSkeleton:
-		return &projSkeletonBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        patterns,
-		}, nil
-	case colexeccmp.LikeSkeletonNegate:
-		return &projNotSkeletonBytesBytesConstOp{
-			projConstOpBase: base,
-			constArg:        patterns,
+			negate:          negate,
 		}, nil
 	case colexeccmp.LikeRegexp:
 		re, err := eval.ConvertLikeToRegexp(ctx, string(patterns[0]), false, '\\')
@@ -115,15 +83,19 @@ func GetLikeProjectionOperator(
 		return &projRegexpBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        re,
+			negate:          negate,
 		}, nil
-	case colexeccmp.LikeRegexpNegate:
-		re, err := eval.ConvertLikeToRegexp(ctx, string(patterns[0]), false, '\\')
-		if err != nil {
-			return nil, err
-		}
-		return &projNotRegexpBytesBytesConstOp{
+	case colexeccmp.LikeSkeleton:
+		return &projSkeletonBytesBytesConstOp{
 			projConstOpBase: base,
-			constArg:        re,
+			constArg:        patterns,
+			negate:          negate,
+		}, nil
+	case colexeccmp.LikeSuffix:
+		return &projSuffixBytesBytesConstOp{
+			projConstOpBase: base,
+			constArg:        pat,
+			negate:          negate,
 		}, nil
 	default:
 		return nil, errors.AssertionFailedf("unsupported like op type %d", likeOpType)