--- a/pkg/sql/colexec/colexecprojconst/like_ops.go
+++ b/pkg/sql/colexec/colexecprojconst/like_ops.go
@@ -31,8 +31,9 @@ func GetLikeProjectionOperator(
 	resultIdx int,
 	pattern string,
 	negate bool,
+	caseInsensitive bool,
 ) (colexecop.Operator, error) {
-	likeOpType, patterns, err := colexeccmp.GetLikeOperatorType(pattern)
+	likeOpType, patterns, err := colexeccmp.GetLikeOperatorType(pattern, caseInsensitive)
 	if err != nil {
 		return nil, err
 	}
@@ -46,13 +47,21 @@ func GetLikeProjectionOperator(
 	}
 	switch likeOpType {
 	case colexeccmp.LikeAlwaysMatch:
-		// Use an empty prefix operator to get correct NULL behavior.
+		// Use an empty prefix operator to get correct NULL behavior. We don't
+		// need to pay attention to the case sensitivity here since the pattern
+		// will always match anyway.
 		return &projPrefixBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        []byte{},
 			negate:          negate,
 		}, nil
 	case colexeccmp.LikeConstant:
+		if caseInsensitive {
+			// We don't have an equivalent projection operator that would
+			// convert the argument to capital letters, so for now we fall back
+			// to the default comparison operator.
+			return nil, errors.New("ILIKE and NOT ILIKE aren't supported with a constant pattern")
+		}
 		if negate {
 			return &projNEBytesBytesConstOp{
 				projConstOpBase: base,
@@ -68,15 +77,17 @@ func GetLikeProjectionOperator(
 			projConstOpBase: base,
 			constArg:        pat,
 			negate:          negate,
+			caseInsensitive: caseInsensitive,
 		}, nil
 	case colexeccmp.LikePrefix:
 		return &projPrefixBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        pat,
 			negate:          negate,
+			caseInsensitive: caseInsensitive,
 		}, nil
 	case colexeccmp.LikeRegexp:
-		re, err := eval.ConvertLikeToRegexp(ctx, string(patterns[0]), false, '\\')
+		re, err := eval.ConvertLikeToRegexp(ctx, string(patterns[0]), caseInsensitive, '\\')
 		if err != nil {
 			return nil, err
 		}
@@ -90,12 +101,14 @@ func GetLikeProjectionOperator(
 			projConstOpBase: base,
 			constArg:        patterns,
 			negate:          negate,
+			caseInsensitive: caseInsensitive,
 		}, nil
 	case colexeccmp.LikeSuffix:
 		return &projSuffixBytesBytesConstOp{
 			projConstOpBase: base,
 			constArg:        pat,
 			negate:          negate,
+			caseInsensitive: caseInsensitive,
 		}, nil
 	default:
 		return nil, errors.AssertionFailedf("unsupported like op type %d", likeOpType)