--- a/pkg/sql/colexec/colexecprojconst/proj_const_ops_tmpl.go
+++ b/pkg/sql/colexec/colexecprojconst/proj_const_ops_tmpl.go
@@ -93,6 +93,9 @@ type _OP_CONST_NAME struct {
 	// {{else}}
 	constArg _R_GO_TYPE
 	// {{end}}
+	// {{if .Negatable}}
+	negate bool
+	// {{end}}
 }
 
 func (p _OP_CONST_NAME) Next() coldata.Batch {
@@ -103,6 +106,14 @@ func (p _OP_CONST_NAME) Next() coldata.Batch {
 	//     variable of type `colexecutils.BinaryOverloadHelper`.
 	// */}}
 	_overloadHelper := p.BinaryOverloadHelper
+	// {{end}}
+	// {{if .Negatable}}
+	// {{/*
+	//     In order to inline the templated code of the LIKE overloads, we need
+	//     to have a `_negate` local variable indicating whether the assignment
+	//     should be negated.
+	// */}}
+	_negate := p.negate
 	// {{end}}
 	batch := p.Input.Next()
 	n := batch.Length()