--- a/pkg/sql/colexec/colexecprojconst/proj_const_ops_tmpl.go
+++ b/pkg/sql/colexec/colexecprojconst/proj_const_ops_tmpl.go
@@ -96,6 +96,9 @@ type _OP_CONST_NAME struct {
 	// {{if .Negatable}}
 	negate bool
 	// {{end}}
+	// {{if .CaseInsensitive}}
+	caseInsensitive bool
+	// {{end}}
 }
 
 func (p _OP_CONST_NAME) Next() coldata.Batch {
@@ -114,6 +117,14 @@ func (p _OP_CONST_NAME) Next() coldata.Batch {
 	//     should be negated.
 	// */}}
 	_negate := p.negate
+	// {{ end }}
+	// {{if .CaseInsensitive}}
+	// {{/*
+	//     In order to inline the templated code of the LIKE overloads, we need
+	//     to have a `_caseInsensitive` local variable indicating whether the
+	//     operator is case insensitive.
+	// */}}
+	_caseInsensitive := p.caseInsensitive
 	// {{end}}
 	batch := p.Input.Next()
 	n := batch.Length()