… pkg/sql/colexec/colexecsel/like_ops_test.go
… pkg/sql/colexec/colexecsel/sel_like_ops.eg.go
… pkg/sql/colexec/colexecsel/selection_ops_tmpl.go
… pkg/sql/colexec/execgen/cmd/execgen/like_ops_gen.go
… pkg/sql/colexec/execgen/cmd/execgen/overloads_base.go
… pkg/sql/opt/exec/execbuilder/testdata/tpch_vec